Remove SceneDelegate to support minimum deployment target of 12.4
