+++ PoliServices/Infrastructure/Extensions/Date+Extension.swift
@@ -1,10 +1,3 @@
-//
-//  String+Date.swift
-//  PoliServices
-//
-//  Created by Personal on 17/01/23.
-//
-
 import Foundation
 
 extension Date {
